requirements.txt
Flask==3.1.2
Pillow==11.3.0
python-dotenv==1.1.1
google-generativeai==0.8.5
gunicorn==23.0.0